pyproject.toml
[tool.poetry]
name = "university"
version = "0.1.0"
description = ""
authors = ["Master yoda"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.1"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
SQLAlchemy = "^1.4.45"
envparse = "^0.2.0"
psycopg2 = "^2.9.6"
asyncpg = "^0.27.0"
alembic = "^1.9.0"
pydantic = {extras = ["email"], version = "^1.10.9"}
pytest = "^7.4.2"
httpx = "^0.25.0"
pytest-asyncio = "^0.21.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
attrs = "^23.1.0"
python-multipart = "^0.0.6"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pre-commit = "^3.5.0"
flake8 = "^6.1.0"
reorder-python-imports = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = ["./src"]
asyncio_mode = "auto"
